feat: add hooks for citation preview, mermaid rendering, and server status
- Introduced `useCitationPreview` hook for managing citation preview state.
- Added `useMermaidRenderer` hook for rendering Mermaid diagrams with error handling and loading states.
- Implemented `useServerStatus` hook to fetch and manage server connection status and tool information.

feat: create error handling and stream processing utilities

- Developed `handleGenkitError` and `handleStreamError` functions for standardized error responses.
- Created `processChatStream` function to handle streaming responses and manage tool invocations.

feat: enhance session storage and SSE utilities

- Updated `JsonSessionStore` to dynamically resolve session directory paths based on runtime environment.
- Added `formatSSE` utility for consistent formatting of Server-Sent Events.

feat: implement chat service and RAG functionality

- Created `streamChatResponse` function for handling chat response streaming.
- Refactored RAG service to support dynamic imports and improved file handling.

feat: improve Mermaid chart utilities and message normalization

- Enhanced `getCleanedMermaidChart` function for better chart processing.
- Added `validateChartCompleteness` function to ensure valid Mermaid diagrams.
- Implemented `copyAsImage` and `copyAsText` functions for exporting diagrams.
- Introduced `normalizeMessageContent` utility for consistent message formatting.
