+++ debug-build.js
@@ -1,0 +1,1 @@
+console.log('[BUILD DEBUG] services-config route analyzing...', typeof window, process.env.NODE_ENV);
