+++ src/services/chatService.ts
@@ -1,0 +1,18 @@
+export async function streamChatResponse(
+  requestBody: any,
+  endpointUrl: string,
+): Promise<Response> {
+  const response = await fetch(endpointUrl, {
+    method: "POST",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(requestBody),
+  });
+
+  if (!response.ok) {
+    throw new Error(`HTTP error! status: ${response.status}`);
+  }
+
+  return response;
+}
