+++ src/utils/message-normalization.ts
@@ -1,0 +1,22 @@
+export function normalizeMessageContent(content: any): string {
+  if (typeof content === "string") {
+    return content;
+  }
+  if (Array.isArray(content)) {
+    return content
+      .map((item) => {
+        if (typeof item === "string") {
+          return item;
+        }
+        if (item && typeof item === "object" && "text" in item) {
+          return item.text;
+        }
+        return JSON.stringify(item);
+      })
+      .join("");
+  }
+  if (content && typeof content === "object" && "text" in content) {
+    return content.text;
+  }
+  return JSON.stringify(content);
+}
